services/job-extractor/requirements.txt
flask==3.1.0
flask-cors==5.0.0
openai==1.59.9
structlog==24.4.0
requests==2.32.3
pytz==2024.2
werkzeug==3.1.3
jinja2==3.1.4
python-dotenv==1.0.1
gunicorn==23.0.0
supabase==2.9.1
beautifulsoup4==4.12.3
PyMuPDF==1.24.12
python-docx==1.1.2